# openapi.yaml
openapi: 3.0.3
info:
  title: Payment Gateway API
  version: 1.0.0
  description: |
    Stripe-style REST API for the in-house payment gateway.

    ## Authentication

    Two parallel auth systems:

    | Use case | Header |
    |----------|--------|
    | Merchant API consumers | `Authorization: Bearer sk_test_xxx_yyy` (or `sk_live_…`) |
    | Dashboard sessions | `Authorization: Bearer <jwt>` (issued by `/auth/login` + `/auth/totp/verify`) |

    Public keys (`pk_test_…` / `pk_live_…`) are restricted to the `tokenize` scope and are safe to ship to the browser.

    ## Idempotency

    Every `POST /v1/payment_intents` request **must** include an `Idempotency-Key` header.
    Replaying with the same `(merchantId, idempotencyKey)` returns the original payment intent.

    ## Pagination

    List endpoints use cursor-based pagination:

    ```
    GET /v1/customers?limit=20&cursor=cus_abc123
    → { data: [...], has_more: true, next_cursor: "cus_xyz789" }
    ```

    ## Errors

    All errors follow a consistent shape:

    ```json
    { "error": { "code": "validation_error", "message": "...", "param": "amount" } }
    ```
  contact:
    name: Payments Engineering
    email: payments@example.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://sandbox.payments.example.com/api/v1
    description: Sandbox
  - url: https://api.payments.example.com/api/v1
    description: Production
...
components:
  securitySchemes:
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: api_key
      description: |
        Merchant API key. Format: `sk_test_xxx_yyy` or `sk_live_xxx_yyy`.
        Public keys (`pk_…`) are restricted to the `tokenize` scope.
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Dashboard session JWT issued by `POST /auth/login` (partial)
        and elevated by `POST /auth/totp/verify` (full).
  parameters:
...
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CursorParam:
      name: cursor
      in: query
      schema:
        type: string
      description: ID of the last item in the previous page
  schemas:
...
    Customer:
      type: object
      properties:
        id: { type: string, format: uuid }
        externalId: { type: string }
        email: { type: string, format: email, nullable: true }
        phone: { type: string, nullable: true }
        name: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...
    PaginatedCustomers:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Customer' }
        has_more: { type: boolean }
        next_cursor:
          type: string
          nullable: true
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Invalid request body or parameters
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
paths:
...
  /merchants/customers:
    get:
      tags: [Merchants]
      summary: List all customers the merchant has interacted with
      description: |
        Returns every customer either owned by this merchant or who has ever
        paid them. Aggregates (`totalSpent`, `totalPayments`, `lastPaymentAt`)
        are scoped to this merchant only — a customer's spend with another
        merchant does not bleed into your numbers. Powers the merchant
        dashboard's "All Customers" panel.
      security: [{ jwtAuth: [] }]
      responses:
        '200':
          description: Customers and aggregates.
          content:
            application/json:
              schema:
                type: object
                properties:
                  customers:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        name: { type: string, nullable: true }
                        email: { type: string, format: email, nullable: true }
                        phone: { type: string, nullable: true }
                        createdAt: { type: string, format: date-time }
                        walletBalance: { type: integer, description: paise }
                        totalSpent: { type: integer, description: paise across SUCCEEDED intents with this merchant }
                        totalPayments: { type: integer, description: count of intents with this merchant (any status) }
                        successfulPayments: { type: integer }
                        paymentMethodCount: { type: integer }
                        lastPaymentAt: { type: string, format: date-time, nullable: true }
                        currency: { type: string, example: INR }
                        isOwnCustomer:
                          type: boolean
                          description: |
                            true if `customer.merchantId === merchant`, false if
                            the customer's home merchant is different but they
                            paid you anyway (e.g. via the public checkout link).
    post:
      tags: [Merchants, Customers]
      summary: Create a customer (dashboard JWT)
      description: |
        Dashboard-friendly equivalent of POST /customers. The merchant JWT
        replaces the customers:write API key. `externalId` is optional —
        when omitted, the server auto-generates one (`cust_<random>`).
      security: [{ jwtAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                externalId: { type: string }
                name: { type: string }
                email: { type: string, format: email }
                phone: { type: string }
      responses:
        '201':
          description: Customer created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  customer: { $ref: '#/components/schemas/Customer' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/Conflict' }
  /merchants/customers/{id}:
    get:
      tags: [Merchants, Customers]
      summary: Customer drill-down (dashboard JWT)
      description: |
        Returns full profile + payment methods + recent intents (scoped to
        this merchant) + wallet balance. Visible if the customer is owned
        by this merchant OR has paid them at least once.
      security: [{ jwtAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Customer detail.
        '404': { $ref: '#/components/responses/NotFound' }
...
  /customers:
    get:
      tags: [Customers]
      summary: List customers (cursor pagination)
      security: [{ apiKeyAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
      responses:
        '200':
          description: Paged customer list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedCustomers' }
    post:
      tags: [Customers]
      summary: Create a customer
      security: [{ apiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [externalId]
              properties:
                externalId: { type: string, minLength: 1, maxLength: 200 }
                email: { type: string, format: email }
                phone: { type: string, example: '+919876543210' }
                name: { type: string, maxLength: 200 }
      responses:
        '201':
          description: Customer created.
        '409': { $ref: '#/components/responses/Conflict' }
  /customers/{id}:
    get:
      tags: [Customers]
      summary: Get a customer by ID (with attached payment methods)
      security: [{ apiKeyAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Customer detail.
...
  /customers/register/public:
    post:
      tags: [Customers]
      summary: Public customer self-registration (no API key)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name]
              properties:
                email: { type: string, format: email }
                name: { type: string }
                phone: { type: string }
      responses:
        '201':
          description: Customer created.
        '200':
          description: Existing customer returned (idempotent).
  /customers/login/public:
    post:
      tags: [Customers]
      summary: Public customer login by email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: Customer found.
        '404': { $ref: '#/components/responses/NotFound' }
  /customers/{id}/methods/{methodId}/public:
    delete:
      tags: [Customers]
      summary: Customer-portal removal of a saved payment method
      description: |
        Auth model is "owner-by-knowledge-of-id" — both the customer ID and
        method ID must match. The same trust model applies to other /public
        flows on this router (the customer portal does not hold a JWT or API
        key; it only knows the customer ID it was logged in with).
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: methodId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Method deleted.
        '404': { $ref: '#/components/responses/NotFound' }
  /customers/{id}/wallet/public:
    get:
      tags: [Wallets]
      summary: Public wallet balance + transaction list
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Wallet snapshot.
  /customers/{id}/wallet/order/public:
    post:
      tags: [Wallets]
      summary: Create a Razorpay Order for wallet top-up
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: integer, minimum: 100 }
      responses:
        '200':
          description: Order ID + Razorpay key id.
  /customers/{id}/wallet/topup/public:
    post:
      tags: [Wallets]
      summary: Confirm Razorpay payment and credit the wallet
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, razorpay_payment_id, razorpay_order_id, razorpay_signature]
              properties:
                amount: { type: integer, minimum: 1 }
                razorpay_payment_id: { type: string }
                razorpay_order_id: { type: string }
                razorpay_signature: { type: string }
      responses:
        '200':
          description: Wallet credited.
  /customers/{id}/wallet/withdraw/public:
    post:
      tags: [Wallets]
      summary: Withdraw from wallet
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Wallet debited.
...
  /customers/{id}/wallet/topup/dummy:
    post:
      tags: [Wallets]
      summary: Public wallet top-up with a simulated outcome (no Razorpay)
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: integer, minimum: 1 }
                outcome:
                  type: string
                  enum: [SUCCESS, FAILURE_DECLINED, FAILURE_REVERTED]
                  default: SUCCESS
                payment_method:
                  type: string
                  default: card
      responses:
        '200':
          description: Outcome applied.
security:
  - apiKeyAuth: []